⚡ — Cache compressed JS and CSS files on disk
JS and CSS files are compressed every time a user asked for them
(unless they are cached by the browser, of course).
This commit caches compressed files on disk (in `.compressed` folder).

The cache folder is removed on startup, when installing or removing plugins
and when admin asks to restart server on /admin/settings.